requirements.txt
pyyaml
numpy
colour
pendulum
twine
cn2an
colorlog
concurrent-log-handler
Deprecated
GitPython